+++ createMat.h
@@ -12,7 +12,7 @@ struct memory {
 };
 
 struct matrix {
-  int					*mat;
+  double				*mat;
   int					lines;
   int					columns;
 };
@@ -26,7 +26,7 @@ int endline,int begincolumn, int endcolumn);
 //Add matrice to memory
 void add_Mat(struct memory *bank, struct matrix *mat, int nbmat);
 
-void print_matrix(int *mat, int lines, int columns);
+void print_matrix(double *mat, int lines, int columns);
 
 void print_all_matrix(struct memory *bank, int size);
